week8/project2/requirements.txt
python-fasthtml>=0.6.0
supabase>=2.0.0
python-dotenv>=1.0.0
